requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
jinja2==3.1.4
notion-client==2.2.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4
cachetools==5.3.3
astral==3.2
aiosmtplib==3.0.1
python-dateutil==2.9.0
slowapi==0.1.9
pytest==8.3.0
httpx==0.27.0
python-dotenv==1.0.1
icalendar==6.1.0
